pyproject.toml
[tool.poetry]
name = "reto-spot"
version = "0.1.0"
description = "challenge spot"
authors = ["Daniel Cabrera <danielmtv6@gmail.com>"]
license = "MIT"
readme = "README.md"


[tool.poetry.dependencies]
python = "^3.9"


[tool.poetry.group.dev.dependencies]
alembic = "1.11.1"
anyio = "3.6.2"
attrs = "23.1.0"
autoflake = "2.1.1"
autopep8 = "2.0.2"
azure-core = "1.26.4"
azure-storage-blob = "12.16.0"
black = "23.3.0"
build = "0.10.0"
cachecontrol = "0.12.11"
certifi = "2023.5.7"
cffi = "1.15.1"
charset-normalizer = "3.1.0"
cleo = "2.0.1"
click = "8.1.3"
coverage = "7.2.5"
crashtest = "0.4.1"
cryptography = "40.0.2"
distlib = "0.3.6"
dulwich = "0.21.5"
exceptiongroup = "1.1.1"
fastapi = "0.95.2"
filelock = "3.12.0"
flake8 = "6.0.0"
greenlet = "2.0.2"
gunicorn = "20.1.0"
h11 = "0.14.0"
html5lib = "1.1"
idna = "3.4"
importlib-metadata = "6.6.0"
iniconfig = "2.0.0"
install = "1.3.5"
installer = "0.7.0"
isort = "5.12.0"
jaraco-classes = "3.2.3"
jeepney = "0.8.0"
jsonschema = "4.17.3"
keyring = "23.13.1"
lockfile = "0.12.2"
mako = "1.2.4"
markupsafe = "2.1.2"
mccabe = "0.7.0"
more-itertools = "9.1.0"
msgpack = "1.0.5"
mypy = "1.3.0"
mypy-extensions = "1.0.0"
packaging = "23.1"
pathspec = "0.11.1"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "3.5.1"
pluggy = "1.0.0"
poetry = "1.5.0"
poetry-core = "1.6.0"
poetry-plugin-export = "1.3.1"
psycopg2-binary = "2.9.6"
ptyprocess = "0.7.0"
pycodestyle = "2.10.0"
pycparser = "2.21"
pydantic = "1.10.7"
pyflakes = "3.0.1"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
pytest = "7.3.1"
pytest-cov = "4.0.0"
pytest-mock = "3.10.0"
rapidfuzz = "2.15.1"
requests = "2.30.0"
requests-toolbelt = "1.0.0"
secretstorage = "3.3.3"
shellingham = "1.5.0.post1"
six = "1.16.0"
sniffio = "1.3.0"
sqlalchemy = "2.0.15"
sqlalchemy-stubs = "0.4"
starlette = "0.27.0"
tomli = "2.0.1"
tomlkit = "0.11.8"
trove-classifiers = "2023.5.2"
typing-extensions = "4.5.0"
urllib3 = "1.26.15"
uvicorn = "0.22.0"
virtualenv = "20.23.0"
webencodings = "0.5.1"
zipp = "3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
